Refactor: Restructure the Project Structure to have modularized design
