+++ app/__init__.py
@@ -2,8 +2,8 @@ from flask import Flask
 from flask_jwt_extended import JWTManager
 from app.config.db import db
 from app.config.config import Config
-from app.routes.auth_routes import auth_bp
-from app.routes.admin_routes import admin_bp
+from app.auth.auth_routes import auth_bp
+from app.admin.admin_routes import admin_bp
 
 
 def create_app():
