+++ app/admin/services/department_service.py
@@ -1,4 +1,4 @@
-from app.repositories.department_repository import find_department_by_name, create_department_record, get_all_departments
+from app.admin.repositories.department_repository import find_department_by_name, create_department_record, get_all_departments
 
 def create_department_service(data):
     name = data.get("name")
